factor out functions sourced from wv_sat_methods.f90
