+++ src/mam4xx/CMakeLists.txt
@@ -22,6 +22,7 @@ install(FILES
         ndrop.hpp
         vehkamaki2002.hpp
         wang2008.hpp
+        wv_sat_methods.hpp
         DESTINATION include/mam4xx)
 
 add_library(mam4xx aero_modes.cpp)
