+++ src/mam4xx/nucleate_ice.hpp
@@ -15,6 +15,7 @@
 #include <mam4xx/conversions.hpp>
 #include <mam4xx/mam4_types.hpp>
 #include <mam4xx/utils.hpp>
+#include <mam4xx/wv_sat_methods.hpp>
 
 namespace mam4 {
 
@@ -49,136 +50,6 @@ clouds, !  following X. Shi et al. ACP (2014).
 !
 !  With help from C. C. Chen and B. Eaton (2014)
 !-------------------------------------------------------------------------------*/
-
-// FIXME from wv_sat_methods.F90
-//  Do we need to create a new files for these functions?
-
-KOKKOS_INLINE_FUNCTION
-Real GoffGratch_svp_water(const Real temperature) {
-  // ! Goff & Gratch (1946)
-  // temperature  ! Temperature in Kelvin
-  // 	es             ! SVP in Pa
-  // ! uncertain below -70 C
-  // BAD CONSTANT
-  // FIXME where should we add this constant
-  // FROM wv_saturation.F90
-  // Boiling point of water at 1 atm (K)
-  // This value is slightly high, but it seems to be the value for the
-  // steam point of water originally (and most frequently) used in the
-  // Goff & Gratch scheme.
-  // FIXME add this constant to haero::Constants
-  const Real tboil = 373.16;
-
-  const Real ten = 10;
-  const Real one = 1;
-  return haero::pow(ten,
-                    -Real(7.90298) * (tboil / temperature - one) +
-                        Real(5.02808) * haero::log10(tboil / temperature) -
-                        Real(1.3816e-7) *
-                            (haero::pow(ten, Real(11.344) *
-                                                 (one - temperature / tboil)) -
-                             one) +
-                        Real(8.1328e-3) *
-                            (haero::pow(ten, -Real(3.49149) *
-                                                 (tboil / temperature - one)) -
-                             one) +
-                        haero::log10(Real(1013.246))) *
-         ten * ten;
-
-} // GoffGratch_svp_water
-
-KOKKOS_INLINE_FUNCTION
-Real GoffGratch_svp_ice(const Real temperature) {
-  // temperature  ! Temperature in Kelvin
-  //  es             ! SVP in Pa
-
-  // ! good down to -100 C
-  const Real h2otrip = haero::Constants::triple_pt_h2o;
-  const Real ten = 10;
-  const Real one = 1;
-
-  return haero::pow(ten,
-                    -Real(9.09718) * (h2otrip / temperature - one) -
-                        Real(3.56654) * haero::log10(h2otrip / temperature) +
-                        Real(0.876793) * (one - temperature / h2otrip) +
-                        haero::log10(6.1071)) *
-         ten * ten;
-
-} // end GoffGratch_svp_ice
-
-// FIXME
-// Compute saturation vapor pressure over water
-KOKKOS_INLINE_FUNCTION
-Real svp_water(const Real Temperature) {
-  // FIXME
-  // ask if we need to implement the other methods to compute svp_water
-  // initial_default_idx = GoffGratch_idx
-  return GoffGratch_svp_water(Temperature);
-}
-
-/*---------------------------------------------------------------------
-! UTILITIES
-!---------------------------------------------------------------------*/
-
-// Get saturation specific humidity given pressure and SVP.
-// Specific humidity is limited to range 0-1.
-KOKKOS_INLINE_FUNCTION
-Real wv_sat_svp_to_qsat(const Real es, const Real p) {
-  // es  ! SVP
-  // p   ! Current pressure.
-  // qs
-  // If pressure is less than SVP, set qs to maximum of 1.
-  // epsilo  ! Ice-water transition range
-  // omeps   ! 1._r8 - epsilo
-  // epsilo       = shr_const_mwwv/shr_const_mwdair   ! ratio of h2o to dry air
-  // molecular weights real(R8),parameter :: SHR_CONST_MWDAIR  = 28.966_R8 !
-  // molecular weight dry air ~ kg/kmole real(R8),parameter :: SHR_CONST_MWWV
-  // = 18.016_R8       ! molecular weight water vapor
-  // FIXME: move these constants to hearo::constants
-  const Real SHR_CONST_MWWV = 18.016;
-  const Real SHR_CONST_MWDAIR = 28.966;
-  const Real epsilo = SHR_CONST_MWWV / SHR_CONST_MWDAIR;
-
-  const Real zero = 0;
-  const Real one = 1;
-  const Real omeps = one - epsilo;
-  Real qs = zero;
-  if ((p - es) <= zero) {
-    qs = one;
-  } else {
-    qs = epsilo * es / (p - omeps * es);
-  }
-  return qs;
-
-} // wv_sat_svp_to_qsat
-
-KOKKOS_INLINE_FUNCTION
-void wv_sat_qsat_water(const Real t, const Real p, Real &es, Real &qs)
-
-{
-  /*------------------------------------------------------------------!
-  ! Purpose:                                                         !
-  !   Calculate SVP over water at a given temperature, and then      !
-  !   calculate and return saturation specific humidity.             !
-  !------------------------------------------------------------------*/
-  // Inputs
-  // t    ! Temperature
-  // p    ! Pressure
-  // Outputs
-  // es  ! Saturation vapor pressure
-  // qs  ! Saturation specific humidity
-
-  es = svp_water(t);
-  qs = wv_sat_svp_to_qsat(es, p);
-  // Ensures returned es is consistent with limiters on qs.
-  es = haero::min(es, p);
-
-} // wv_sat_qsat_water
-
-KOKKOS_INLINE_FUNCTION
-Real svp_ice(const Real Temperature) { return GoffGratch_svp_ice(Temperature); }
-
-// From here all functions are from nucleate_ice
 
 KOKKOS_INLINE_FUNCTION
 void calculate_regm_nucleati(const Real w_vlc, const Real Na, Real &regm) {
@@ -504,7 +375,7 @@ public:
             Real es = zero;
             Real qs = zero;
 
-            nucleate_ice::wv_sat_qsat_water(temp, pmid, es, qs);
+            wv_sat_methods::wv_sat_qsat_water(temp, pmid, es, qs);
             const Real relhum = qv / qs;
 
             // Real relhum =
@@ -634,8 +505,8 @@ public:
     const Real num_threshold = 1.0e-10;
 
     if (so4_num >= num_threshold && dst3_num >= num_threshold && cldn > zero) {
-      if ((tc <= Real(-35.0)) && (relhum * nucleate_ice::svp_water(tair) /
-                                      nucleate_ice::svp_ice(tair) * subgrid >=
+      if ((tc <= Real(-35.0)) && (relhum * wv_sat_methods::svp_water(tair) /
+                                      wv_sat_methods::svp_ice(tair) * subgrid >=
                                   Real(1.2))) {
         //! use higher RHi threshold
         nucleate_ice::calculate_regm_nucleati(wbar, dst3_num, regm);
